+++ inc/customizer/controls/color/class-control-color.php
@@ -83,7 +83,9 @@ class OceanWP_Customizer_Color_Control extends WP_Customize_Control {
 			<# if ( data.description ) { #>
 				<span class="description customize-control-description">{{{ data.description }}}</span>
 			<# } #>
-			<input class="alpha-color-control" type="text"  value="{{ data.value }}" data-show-opacity="{{ data.show_opacity }}" data-palette="{{ data.palette }}" data-default-color="{{ data.default }}" {{{ data.link }}} />
+			<div>
+				<input class="alpha-color-control" type="text"  value="{{ data.value }}" data-show-opacity="{{ data.show_opacity }}" data-palette="{{ data.palette }}" data-default-color="{{ data.default }}" {{{ data.link }}} />
+			</div>
 		</label>
 		<?php
 	}