+++ inc/helpers.php
@@ -1237,6 +1237,55 @@ if ( ! function_exists( 'oceanwp_medium_header_elements' ) ) {
 }
 
 /**
+ * Custom nav template
+ *
+ * @since 1.4.7
+ */
+if ( ! function_exists( 'oceanwp_custom_nav_template' ) ) {
+
+	function oceanwp_custom_nav_template() {
+
+		// Get template from customizer setting
+		$template = get_theme_mod( 'ocean_custom_nav_template' );
+
+		// Apply filters and return
+		return apply_filters( 'ocean_custom_nav_template', $template );
+
+	}
+
+}
+
+/**
+ * Returns header template content
+ *
+ * @since 1.4.7
+ */
+if ( ! function_exists( 'oceanwp_nav_template_content' ) ) {
+
+	function oceanwp_nav_template_content() {
+
+		// Get the template ID
+		$content = oceanwp_custom_nav_template();
+
+		// Get template content
+		if ( ! empty( $content ) ) {
+
+			$template = get_post( $content );
+
+			if ( $template && ! is_wp_error( $template ) ) {
+				$content = $template->post_content;
+			}
+
+		}
+
+		// Apply filters and return content
+		return apply_filters( 'oceanwp_nav_template_content', $content );
+
+	}
+
+}
+
+/**
  * Returns correct menu classes
  *
  * @since 1.0.0
@@ -3490,72 +3539,6 @@ if ( ! function_exists( 'oceanwp_minify_css' ) ) {
 		// Return minified CSS
 		return $css;
 		
-	}
-
-}
-
-/**
- * Minify JS
- *
- * @since 1.1.0
- */
-if ( ! function_exists( 'oceanwp_minify_js' ) ) {
-
-	function oceanwp_minify_js( $js = '' ) {
-
-		// Return if no JS
-		if ( ! $js ) return;
-
-		if ( OCEAN_EXTRA_ACTIVE
-			&& class_exists( 'Ocean_Extra_JSMin' ) ) {
-
-			$script = Ocean_Extra_JSMin::minify( $js );
-
-		} else {
-
-			$replace = array(
-				'#\'([^\n\']*?)/\*([^\n\']*)\'#' 	=> "'\1/'+\'\'+'*\2'", 	// remove comments from ' strings
-				'#\"([^\n\"]*?)/\*([^\n\"]*)\"#' 	=> '"\1/"+\'\'+"*\2"', 	// remove comments from " strings
-				'#/\*.*?\*/#s'            			=> "",      			// strip C style comments
-				'#[\r\n]+#'               			=> "\n",    			// remove blank lines and \r's
-				'#\n([ \t]*//.*?\n)*#s'   			=> "\n",    			// strip line comments (whole line only)
-				'#([^\\])//([^\'"\n]*)\n#s' 		=> "\\1\n", 			// strip line comments
-				'#\n\s+#'                 			=> "\n",    			// strip excess whitespace
-				'#\s+\n#'                 			=> "\n",    			// strip excess whitespace
-				'#(//[^\n]*\n)#s'         			=> "\\1\n", 			// extra line feed after any comments left
-				'#/([\'"])\+\'\'\+([\'"])\*#' 		=> "/*" 				// restore comments in strings
-			);
-
-			$search = array_keys( $replace );
-			$script = preg_replace( $search, $replace, $js );
-
-			$replace = array(
-				"&&\n" => "&&",
-				"||\n" => "||",
-				"(\n"  => "(",
-				")\n"  => ")",
-				"[\n"  => "[",
-				"]\n"  => "]",
-				"+\n"  => "+",
-				",\n"  => ",",
-				"?\n"  => "?",
-				":\n"  => ":",
-				";\n"  => ";",
-				"{\n"  => "{",
-				"\n]"  => "]",
-				"\n)"  => ")",
-				"\n}"  => "}",
-				"\n\n" => "\n"
-			);
-
-			$search = array_keys( $replace );
-			$script = str_replace( $search, $replace, $script );
-
-		}
-
-		// Return minified JS
-		return trim( $script );
-
 	}
 
 }